experimentoLatencia/services/products/requirements.txt
Flask
Flask-Caching
redis
pandas
psycopg2-binary
gunicorn